BC-745 Integrate codeowners with coverage
has #comment: WRT python modules, glob to all the .py files under the
dir, because we have mixed in .html and such that exist underneath a
module, but don't actually have python code owned by it.
A first pass at a circle configuration for precommit.  It had tests, but
nothing to run them before.
After adding the circle config, fix one test and fix a bunch of warnings
elsewhere.